+++ .gitignore
@@ -1,0 +1,25 @@
+# Ignore Bin and Debug folders
+[Bb]in/
+[Dd]ebug/
+
+# Other common .gitignore entries for C# projects
+*.user
+*.suo
+*.userosscache
+*.sln.docstates
+*.vs/
+.vscode/
+*.csproj.user
+*.csproj.vspscc
+*.csproj.vspscc
+*.dbmdl
+*.jfm
+*.pdb
+*.opendb
+*.suo
+*.sln
+*.sln.docstates
+*.user
+*.userosscache
+*.vs/
+.vscode/
