+++ Models/ClockInOutResponse.cs
@@ -1,0 +1,5 @@
+public class ClockInOutResponse
+{
+    public bool Success { get; set; }
+    public string Message { get; set; }
+}
