+++ Models/LoginDto.cs
@@ -1,6 +1,6 @@
 public class LoginDto
 {
-    // public string Username { get; set; }
+    
     public string Password { get; set; }
     public int employeeNo { get; set; }
 }
